+++ src/freeroute/designformats/specctra/DsnFile.java
@@ -313,7 +313,7 @@ public class DsnFile
             Object next_token = p_scanner.next_token();
             if (next_token instanceof Integer)
             {
-                value = ((Integer) next_token).intValue();
+                value = (int) next_token;
             }
             else
             {
@@ -343,11 +343,11 @@ public class DsnFile
             Object next_token = p_scanner.next_token();
             if (next_token instanceof Double)
             {
-                value = ((Double) next_token).doubleValue();
+                value = (double) next_token;
             }
             else if (next_token instanceof Integer)
             {
-                value = ((Integer) next_token).intValue();
+                value = (int) next_token;
             }
             else
             {
