+++ src/freeroute/designformats/specctra/SessionToEagle.java
@@ -502,11 +502,11 @@ public class SessionToEagle extends javax.swing.JFrame
             next_token = this.scanner.next_token();
             if (next_token instanceof Double)
             {
-                location[i] = ((Double) next_token).doubleValue();
+                location[i] = (double) next_token;
             }
             else if (next_token instanceof Integer)
             {
-                location[i] = ((Integer) next_token).intValue();
+                location[i] = (int) next_token;
             }
             else
             {
